% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamsList.R
\name{as.list.ParamsList}
\alias{as.list.ParamsList}
\title{Convert ParamsList to list}
\usage{
\method{as.list}{ParamsList}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ParamsList}.}

\item{...}{Additional arguments (not used).}
}
\value{
A name sorted list representing the ParamsList object.
}
\description{
Converts the current state of the ParamsList object to a list.
}
\seealso{
Other ParamsList-methods: 
\code{\link{[.ParamsList}()},
\code{\link{[<-.ParamsList}()},
\code{\link{[[.ParamsList}()},
\code{\link{[[<-.ParamsList}()},
\code{\link{as.character.ParamsList}()},
\code{\link{names.ParamsList}()},
\code{\link{print.ParamsList}()},
\code{\link{sort.ParamsList}()}
}
\concept{ParamsList-methods}
