% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_url}
\alias{base_url}
\title{Base url of the database API}
\format{
An object of class \code{character} of length 1.
}
\usage{
base_url
}
\description{
Base url of the database API
}
\keyword{datasets}
