% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{clear}
\alias{clear}
\title{Generic clear function}
\usage{
clear(object, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{x}{An object.}
}
\description{
This is the generic clear function.
}
