% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{database_info}
\alias{database_info}
\title{Urls and other information for accessing the database}
\format{
\subsection{\code{database_info}}{

A list of list
\describe{
\item{endpoints}{endpoint addresses for the various views in the db.
These are intended to be appended to the URL of the database itself}
}
}
}
\usage{
database_info
}
\description{
Urls and other information for accessing the database
}
\keyword{datasets}
