% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSourceAPI.R
\name{DataSourceAPI}
\alias{DataSourceAPI}
\title{DataSourceAPI R6 Class}
\value{
A DataSourceAPI object
}
\description{
An R6 class to interact with the DataSourceAPI endpoint.
}
\seealso{
Other API: 
\code{\link{AbstractAPI}},
\code{\link{AbstractRecordsAndFilesAPI}},
\code{\link{AbstractRecordsOnlyAPI}},
\code{\link{BindingManualQCAPI}},
\code{\link{ExpressionAPI}},
\code{\link{ExpressionManualQCAPI}},
\code{\link{FileFormatAPI}},
\code{\link{GenomicFeatureAPI}},
\code{\link{PromoterSetSigAPI}},
\code{\link{RankResponseAPI}},
\code{\link{RegulatorAPI}}
}
\concept{API}
\section{Super classes}{
\code{yeastCCDBShiny::AbstractAPI} -> \code{base::AbstractRecordsOnlyAPI} -> \code{DataSourceAPI}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataSourceAPI-new}{\code{DataSourceAPI$new()}}
\item \href{#method-DataSourceAPI-read}{\code{DataSourceAPI$read()}}
\item \href{#method-DataSourceAPI-clone}{\code{DataSourceAPI$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="yeastCCDBShiny" data-topic="AbstractAPI" data-id="create"><a href='../../yeastCCDBShiny/html/AbstractAPI.html#method-AbstractAPI-create'><code>yeastCCDBShiny::AbstractAPI$create()</code></a></span></li>
<li><span class="pkg-link" data-pkg="yeastCCDBShiny" data-topic="AbstractAPI" data-id="delete"><a href='../../yeastCCDBShiny/html/AbstractAPI.html#method-AbstractAPI-delete'><code>yeastCCDBShiny::AbstractAPI$delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="yeastCCDBShiny" data-topic="AbstractAPI" data-id="pop_params"><a href='../../yeastCCDBShiny/html/AbstractAPI.html#method-AbstractAPI-pop_params'><code>yeastCCDBShiny::AbstractAPI$pop_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="yeastCCDBShiny" data-topic="AbstractAPI" data-id="push_params"><a href='../../yeastCCDBShiny/html/AbstractAPI.html#method-AbstractAPI-push_params'><code>yeastCCDBShiny::AbstractAPI$push_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="yeastCCDBShiny" data-topic="AbstractAPI" data-id="update"><a href='../../yeastCCDBShiny/html/AbstractAPI.html#method-AbstractAPI-update'><code>yeastCCDBShiny::AbstractAPI$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSourceAPI-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataSourceAPI-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceAPI$new(url = Sys.getenv("DATASOURCE_URL"), ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSourceAPI-read"></a>}}
\if{latex}{\out{\hypertarget{method-DataSourceAPI-read}{}}}
\subsection{Method \code{read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceAPI$read(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSourceAPI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataSourceAPI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceAPI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
