% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankResponseAPI.R
\name{.extract_rr_tarball_csv}
\alias{.extract_rr_tarball_csv}
\title{Read in a CSV File from a Rank Response Tarball}
\usage{
.extract_rr_tarball_csv(tar_file, filename)
}
\arguments{
\item{tar_file}{A tarball file containing, at this point, only one
CSV file}

\item{filename}{The name of the CSV file to read from the tarball}
}
\value{
A dataframe with the contents of the specified
CSV file from the tarball
}
\description{
Read in a CSV File from a Rank Response Tarball
}
