% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankResponseAPI.R
\name{.extract_rr_tarball_metadata}
\alias{.extract_rr_tarball_metadata}
\title{Read in the Metadata from a Rank Response Tarball}
\usage{
.extract_rr_tarball_metadata(tar_file, metadata_filename = "metadata.json")
}
\arguments{
\item{tar_file}{A tarball file containing, at this point, only one
CSV file}
}
\value{
A dataframe of metadata with information on the contents of the
Rank Response tarball
}
\description{
Read in the Metadata from a Rank Response Tarball
}
