% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicFeatureAPI.R
\name{GenomicFeatureAPI}
\alias{GenomicFeatureAPI}
\title{GenomicFeatureAPI R6 Class}
\value{
A GenomicFeatureAPI object
}
\description{
An R6 class to interact with the GenomicFeatureAPI endpoint.
}
\seealso{
Other API: 
\code{\link{AbstractAPI}},
\code{\link{AbstractRecordsAndFilesAPI}},
\code{\link{AbstractRecordsOnlyAPI}},
\code{\link{BindingManualQCAPI}},
\code{\link{DataSourceAPI}},
\code{\link{ExpressionAPI}},
\code{\link{ExpressionManualQCAPI}},
\code{\link{FileFormatAPI}},
\code{\link{PromoterSetSigAPI}},
\code{\link{RankResponseAPI}},
\code{\link{RegulatorAPI}}
}
\concept{API}
\section{Super classes}{
\code{yeastCCDBShiny::AbstractAPI} -> \code{base::AbstractRecordsOnlyAPI} -> \code{GenomicFeatureAPI}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GenomicFeatureAPI-new}{\code{GenomicFeatureAPI$new()}}
\item \href{#method-GenomicFeatureAPI-clone}{\code{GenomicFeatureAPI$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="yeastCCDBShiny" data-topic="AbstractAPI" data-id="create"><a href='../../yeastCCDBShiny/html/AbstractAPI.html#method-AbstractAPI-create'><code>yeastCCDBShiny::AbstractAPI$create()</code></a></span></li>
<li><span class="pkg-link" data-pkg="yeastCCDBShiny" data-topic="AbstractAPI" data-id="delete"><a href='../../yeastCCDBShiny/html/AbstractAPI.html#method-AbstractAPI-delete'><code>yeastCCDBShiny::AbstractAPI$delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="yeastCCDBShiny" data-topic="AbstractAPI" data-id="pop_params"><a href='../../yeastCCDBShiny/html/AbstractAPI.html#method-AbstractAPI-pop_params'><code>yeastCCDBShiny::AbstractAPI$pop_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="yeastCCDBShiny" data-topic="AbstractAPI" data-id="push_params"><a href='../../yeastCCDBShiny/html/AbstractAPI.html#method-AbstractAPI-push_params'><code>yeastCCDBShiny::AbstractAPI$push_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="yeastCCDBShiny" data-topic="AbstractAPI" data-id="update"><a href='../../yeastCCDBShiny/html/AbstractAPI.html#method-AbstractAPI-update'><code>yeastCCDBShiny::AbstractAPI$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="base" data-topic="AbstractRecordsOnlyAPI" data-id="read"><a href='../../base/html/AbstractRecordsOnlyAPI.html#method-AbstractRecordsOnlyAPI-read'><code>base::AbstractRecordsOnlyAPI$read()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomicFeatureAPI-new"></a>}}
\if{latex}{\out{\hypertarget{method-GenomicFeatureAPI-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomicFeatureAPI$new(url = Sys.getenv("GENOMICFEATURE_URL"), ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomicFeatureAPI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GenomicFeatureAPI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomicFeatureAPI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
