% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOModule.R
\name{goServer}
\alias{goServer}
\title{GO Module Server}
\usage{
goServer(id, rank_response_df, rank_threshold)
}
\arguments{
\item{id}{A string that gives the namespace for the module's input and
output.}

\item{rank_response_df}{A reactive expression that returns a data frame
containing the rank response.}

\item{rank_threshold}{A reactive expression that returns a numeric value
representing the rank threshold.}
}
\description{
This function defines the server-side functionality of the GO Module.
The GO Module provides the user with two clickable links that lead to
enrichment results for "Kemmeren TFKO" and "McIsaac ZEV".
}
\examples{
\dontrun{
rank_response_df_reactive <- shiny::reactive({
  yeastCCDBShiny::rank_response_df
})

rank_threshold <- shiny::reactive({
  25
})

goServer("go_test",
  rank_response_df = rank_response_df_reactive,
  rank_threshold = rank_threshold
)
}

}
\seealso{
\code{\link{goUI}} for the corresponding UI function.

\code{\link{testGoModule}} for a function that tests the GO Module.
}
