% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOModule.R
\name{goUI}
\alias{goUI}
\title{GO Module User Interface}
\usage{
goUI(id)
}
\arguments{
\item{id}{A string that gives the namespace for the module's input and
output.}
}
\value{
A Shiny tagList containing the UI elements for the module.
}
\description{
This function defines the user interface of the GO Module.
The GO Module presents the user with two clickable buttons leading to
enrichment results for "Kemmeren TFKO" and "McIsaac ZEV".
}
\examples{
\dontrun{
ui <- shiny::fluidPage(
  goUI("go_test")
)
}

}
\seealso{
\code{\link{goServer}} for the corresponding server function.

\code{\link{testGoModule}} for a function that tests the GO Module.
}
