% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hap5_rr_data}
\alias{hap5_rr_data}
\title{Rank Response Data for TF Hap5}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 85707 rows and 17 columns.
}
\usage{
hap5_rr_data
}
\description{
Representative dataframe to examine rank response
}
\keyword{datasets}
