% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_memory_get}
\alias{in_memory_get}
\title{A helper function for httr::GET}
\usage{
in_memory_get(url, headers, params = list())
}
\arguments{
\item{url}{the full url to the API endpoint}

\item{headers}{the headers to include in the request. Should be an
httr::headers object}

\item{params}{the query parameters to include in the request. As a list
where the name is the parameter name and the value is the parameter
value}
}
\description{
This function is a wrapper around httr::GET that writes the response to
memory
}
