% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenAuthModule.R
\name{login_modal}
\alias{login_modal}
\title{Login Modal}
\usage{
login_modal(id)
}
\arguments{
\item{id}{Element ID}
}
\description{
A modal which has a field for username and password, login
and cancel buttons
}
\seealso{
\code{tokenAuthUI}, \code{tokenAuthServer}, \code{tokenAuthModule}
}
