% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login_server}
\alias{login_server}
\title{FUNCTION_TITLE}
\usage{
login_server(id)
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[shiny]{observeEvent}}
\code{\link[httr]{POST}}, \code{\link[httr]{content_type}}
}
