% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login_ui}
\alias{login_ui}
\title{FUNCTION_TITLE}
\usage{
login_ui(id)
}
\arguments{
\item{id}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[shiny]{fluidPage}}, \code{\link[shiny]{textInput}}, \code{\link[shiny]{passwordInput}}, \code{\link[shiny]{actionButton}}
}
