% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_utils.R
\name{multi_query_enrichment_links}
\alias{multi_query_enrichment_links}
\title{Perform Multiple Query Enrichment on the Rank Response Data and Return Links}
\usage{
multi_query_enrichment_links(rank_response_df, rank_threshold)
}
\arguments{
\item{rank_response_df}{A dataframe with rank and target_locus_tag data.}

\item{rank_threshold}{A numeric value to filter rank.}
}
\value{
A named list where the names are source expressions and the values
are short links to gprofiler enrichment results.
}
\description{
This function performs multiple query enrichment using gprofiler and returns
the short links for the enrichment results. The multi_query is performed
separately for each expression source in the data
}
\examples{
df <- tibble::tibble(
  rank = rep(seq(5, 20, 5), each = 5),
  target_locus_tag =
    rep(list(sample(c(
      "YAL001C", "YAL002W", "YAL003W",
      "YAL005C", "YAL007C", "YAL008W",
      "YAL009W", "YAL010C", "YAL011W"
    ), 5)), 20),
  experiment = rep(c("expr1", "expr2"), 10),
  source_expr = rep(c("experiment1", "experiment2"), each = 10)
)

}
