% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamsList.R
\name{names.ParamsList}
\alias{names.ParamsList}
\title{return the names of the parameters}
\usage{
\method{names}{ParamsList}(object)
}
\arguments{
\item{object}{An object of class \code{ParamsList}.}
}
\value{
A character vector of the sorted names of the parameters in the
same order as the input \code{object}
}
\description{
return the names of the parameters
}
\seealso{
Other ParamsList-methods: 
\code{\link{ParamsList}()},
\code{\link{as.character.ParamsList}()},
\code{\link{as.list.ParamsList}()}
}
\concept{ParamsList-methods}
