% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{on_disk_get}
\alias{on_disk_get}
\title{On Disk GET}
\usage{
on_disk_get(url, headers, tar_file, params = list())
}
\arguments{
\item{url}{the URL to GET}

\item{headers}{the headers to include in the request}

\item{tar_file}{the file to write the response to}

\item{params}{the query parameters to include in the request}
}
\value{
the response object
}
\description{
This function is a wrapper around httr::GET that writes the response to disk
}
\note{
this will stop/raise an error if the tar_file directory doesn't exist,
or the GET request fails
}
