% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamsList.R
\name{ParamsList}
\alias{ParamsList}
\title{ParamsList S3 Class}
\usage{
ParamsList(params = list())
}
\arguments{
\item{params}{A list of parameters to initialize the ParamsList object.}
}
\value{
An object of class \code{ParamsList}, sorted by name.
}
\description{
An S3 class to manage a list of parameters with methods to print and clear parameters.
}
\examples{
# Create an instance of ParamsList
params <- ParamsList(list(a = 1, b = 2))
print(params)

# Use [ and [[ to extract elements
print(params['a'])
print(params[['a']])

# Use [<- and [[<- to assign elements
params['c'] <- 3
print(params)

params[c('d', 'e', 'f')] <- list(4, 5, 6)
print(params)

params[['g']] <- 7
print(params)
}
