% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcTableModule.R
\name{qcTableServer}
\alias{qcTableServer}
\title{QC Table Server}
\usage{
qcTableServer(id, tf_id, token)
}
\arguments{
\item{id}{element ID}
}
\description{
Server component of the QC table module
}
\seealso{
\code{qcTableUI}, \code{qcTableModule}
}
