% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcTableModule.R
\name{qcTableUI}
\alias{qcTableUI}
\title{QC table UI component}
\usage{
qcTableUI(id)
}
\arguments{
\item{id}{element ID}
}
\description{
The UI components of the QC table module
}
\seealso{
\code{qcTableServer}, \code{qcTableModule}
}
