% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rank_response_plot}
\alias{rank_response_plot}
\title{Rank Response Plot Example}
\format{
An object of class \code{gg} (inherits from \code{ggplot}) of length 9.
}
\usage{
rank_response_plot
}
\description{
Representative rank response plot. Data from Hap5
}
\keyword{datasets}
