% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankResponsePlotModule.R
\name{rankResponsePlotServer}
\alias{rankResponsePlotServer}
\title{Rank Response Plot Module Server}
\usage{
rankResponsePlotServer(
  id,
  tf_id,
  tf_gene,
  table_row_click,
  token,
  test = FALSE
)
}
\arguments{
\item{id}{Element ID. see \url{https://mastering-shiny.org/scaling-modules.html}}

\item{tf_id}{Reactive Character. The transcription factor ID from the database
backend -- this is the ID of that gene in the genes table. Though this is
an integer, it is passed as a Character vector, eg "1234".}

\item{tf_gene}{Reactive Character. The common name of the gene, eg 'EDS1'}

\item{table_row_click}{Reactive. This reactive is meant to facilitate actions
on the plot, eg highlighting, when the user clicks a given row in the QC
table}

\item{token}{Reactive Character. User authorization token}

\item{test}{Reactive Boolean. Set to TRUE to pull data from the package
resources rather than from the database}

\item{background_source}{Reactive Character. The background against which the hops'
significance is calculated. Suggested default is 'adh1'}

\item{promoter_source}{Reactive Character. The promoter regions definition source
against which the hops' significance is calculated. Suggested default
is 'yiming'}

\item{normalize}{Reactive Boolean. Set to TRUE to normalize the number of responsive
genes across (gene expression) data sets. Suggested default is False}

\item{bin_size}{Reactive Integer. The rank response bin size. Suggested default is 5}

\item{qvalue_thres}{Reactive Decimal. This will control the vertical line which
partitions the rank response plot between values above and below a given
qvalue threshold. See \url{https://github.com/StoreyLab/qvalue}}

\item{confidence_intervals}{Reactive Boolean. Set to TRUE to display the range of
values a given rank response bin may take with a chance of 95\% or greater}
}
\value{
The module server function, but otherwise no explicit reactive
vlaues are returned
}
\description{
Server for the rank response plot module
}
\seealso{
\code{\link{rankResponsePlotUI}}, \code{\link{testRankResponsePlotModule}}
}
