% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankResponsePlotModule.R
\name{rankResponsePlotUI}
\alias{rankResponsePlotUI}
\title{Rank Response Plot Module UI}
\usage{
rankResponsePlotUI(id)
}
\arguments{
\item{id}{Element ID. see \url{https://mastering-shiny.org/scaling-modules.html}}
}
\value{
the module UI
}
\description{
A UI object which contains the plotly output for a rank
response plot
}
\seealso{
\code{rankResponsePlotServer}, \code{testRankResponsePlotModule}
}
