% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_rank_response_data.R
\name{retrieve_rank_response_data}
\alias{retrieve_rank_response_data}
\title{Retrieve and process rank response data from a remote server.}
\usage{
retrieve_rank_response_data(
  tf_gene,
  tf_id,
  hops_source,
  background_source,
  promoter_source,
  token
)
}
\arguments{
\item{tf_gene}{Character. The name of the transcription factor gene.}

\item{tf_id}{Numeric. The id of the transcription factor.}

\item{hops_source}{Character. The source (pipeline) which produced the
qbed files}

\item{background_source}{Character. The source of the background dataset.}

\item{promoter_source}{Character. The source of the promoter dataset.}

\item{token}{Character. The authorization token for the API request.}
}
\value{
ggplot. A plot representing the rank response of the
transcription factor.
}
\description{
This function fetches data related to a specific transcription factor from
a remote server, processes the data, and returns a ggplot object
representing the rank response of the transcription factor.
}
\examples{
\dontrun{
retrieve_rank_response_data(
  tf_gene = "my_gene", tf_id = 123,
  background_source = "source1",
  promoter_source = "source2",
  token = "my_token"
)
}

}
