% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamsList.R
\name{sort.ParamsList}
\alias{sort.ParamsList}
\title{Sort a ParamsList by name}
\usage{
\method{sort}{ParamsList}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{ParamsList}}

\item{decreasing}{A logical value indicating whether to sort
in decreasing order}
}
\value{
A ParamsList object with the parameters sorted by name
}
\description{
Sort a ParamsList by name
}
\seealso{
Other ParamsList-methods: 
\code{\link{[.ParamsList}()},
\code{\link{[<-.ParamsList}()},
\code{\link{[[.ParamsList}()},
\code{\link{[[<-.ParamsList}()},
\code{\link{as.character.ParamsList}()},
\code{\link{as.list.ParamsList}()},
\code{\link{names.ParamsList}()},
\code{\link{print.ParamsList}()}
}
\concept{ParamsList-methods}
