% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamsList.R
\name{[[<-.ParamsList}
\alias{[[<-.ParamsList}
\title{Assign element to ParamsList using \verb{[[<-}}
\usage{
\method{[[}{ParamsList}(x, i) <- value
}
\arguments{
\item{x}{An object of class \code{ParamsList}.}

\item{i}{The name of the parameter to assign.}

\item{value}{The value to assign to the parameter.}
}
\value{
The modified ParamsList object.
}
\description{
Assign element to ParamsList using \verb{[[<-}
}
