% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamsList.R
\name{[<-.ParamsList}
\alias{[<-.ParamsList}
\title{Assign element(s) to ParamsList using \verb{[<-}}
\usage{
\method{[}{ParamsList}(x, i) <- value
}
\arguments{
\item{x}{An object of class \code{ParamsList}.}

\item{i}{The name(s) of the parameter(s) to assign.}

\item{value}{The value(s) to assign to the parameter(s).}
}
\value{
The modified ParamsList object.
}
\description{
Assign element(s) to ParamsList using \verb{[<-}
}
