% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_detail.R
\name{table_detail_server}
\alias{table_detail_server}
\title{FUNCTION_TITLE}
\usage{
table_detail_server(id, db_tablename)
}
\arguments{
\item{id}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[shiny]{moduleServer}}, \code{\link[shiny]{reactive}}, \code{\link[shiny]{req}}, \code{\link[shiny]{validate}}, \code{\link[shiny]{tableOutput}}
\code{\link[tools]{fileutils}}
\code{\link[vroom]{vroom}}
}
