% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_detail.R
\name{table_detail_ui}
\alias{table_detail_ui}
\title{FUNCTION_TITLE}
\usage{
table_detail_ui(id)
}
\arguments{
\item{id}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[shiny]{fluidPage}}, \code{\link[shiny]{fileInput}}, \code{\link[shiny]{tableOutput}}
}
