% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOModule.R
\name{testGoModule}
\alias{testGoModule}
\title{GO Module Test App}
\usage{
testGoModule()
}
\value{
None
}
\description{
A function to launch an app which displays only the GO module UI
and demonstrates the server functionality
}
\examples{
\dontrun{
if (interactive()) {
  testGoModule()
}
}
}
