% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcTableModule.R
\name{testQcTableModule}
\alias{testQcTableModule}
\title{QC Table Module Tester}
\usage{
testQcTableModule()
}
\description{
A minimal test of the QC Example of the QC Table Module
}
\seealso{
\code{qcTableServer}, \code{qcTableModule}
}
