% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankResponsePlotModule.R
\name{testRankResponsePlotModule}
\alias{testRankResponsePlotModule}
\title{Rank Response Plot Module Test App}
\usage{
testRankResponsePlotModule()
}
\value{
None
}
\description{
A function to launch an app which displays only the rank
response module UI and demonstrates the server functionality
}
\examples{
\dontrun{
if (interactive()) {
  testRankResponsePlotModule()
}
}
}
