% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_rank_response_data.R
\name{testRetriveRankResponseData}
\alias{testRetriveRankResponseData}
\title{Test Retrieval of Rank Response Data}
\usage{
testRetriveRankResponseData()
}
\value{
This function does not return a value; it runs a Shiny application.
}
\description{
A function that launches a Shiny application to test and visualize
the retrieval of rank response data.
}
\details{
The Shiny application provides an interface for entering various parameters
related to the transcription factor and data sources, and displays the rank response
plot based on the entered parameters.
}
\examples{
\dontrun{
if (interactive()) {
  # Launch the Shiny application
  testRetriveRankResponseData()
}
}
}
\seealso{
\code{generate_plot_data}, \code{retrieve_rank_response_data}
}
