% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfSelectorModule.R
\name{testTfSelectorModule}
\alias{testTfSelectorModule}
\title{Test TF Selector Module}
\usage{
testTfSelectorModule()
}
\description{
A function which launches an app of only this module
}
\seealso{
\code{\link{tfSelectorUI}}, \code{\link{tfSelectorServer}}
}
