% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfSelectorModule.R
\name{tfSelectorServer}
\alias{tfSelectorServer}
\title{TF Selector Server}
\usage{
tfSelectorServer(id, token, selected_tf)
}
\arguments{
\item{id}{Element ID}

\item{token}{ReactiveValues. keys \code{username} and \code{token}}

\item{selected_tf}{ReactiveValues. keys \code{tf_id} and \code{tf_gene}}
}
\value{
None, but the \code{selected_tf} reactive from the outerscope will
trigger observeEvent in the outer scope.
}
\description{
Tf Selector Module Server
}
\seealso{
\code{\link{tfSelectorServer}, \code{tfSelectorModule}
}
