% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfSelectorModule.R
\name{tfSelectorUI}
\alias{tfSelectorUI}
\title{TF Selector UI}
\usage{
tfSelectorUI(id)
}
\arguments{
\item{id}{Character. Element ID}
}
\description{
UI component of the TF selector module
}
\seealso{
\code{\link{tfSelectorServer}, \code{tfSelectorModule}
}
