% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenAuthModule.R
\name{tokenAuthModule}
\alias{tokenAuthModule}
\title{Token Authentication Module}
\usage{
tokenAuthModule()
}
\description{
A login button which activates a modal to perform user
authentication
}
\seealso{
\code{tokenAuthUI}, \code{tokenAuthModule}
}
