% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenAuthModule.R
\name{tokenAuthServer}
\alias{tokenAuthServer}
\title{Token Authentication Server}
\usage{
tokenAuthServer(id)
}
\arguments{
\item{id}{Element ID}
}
\value{
ReactiveValues. keys \code{username} and \code{token}
}
\description{
Server function of the Token Authentication Module
}
\seealso{
\code{tokenAuthUI}, \code{tokenAuthModule}
\code{\link[labretriever]{get_user_auth_token}}, \code{\link[labretriever]{create_url}}
\code{\link[yeastCCDBShiny]{base_url}}
}
