% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenAuthModule.R
\name{tokenAuthUI}
\alias{tokenAuthUI}
\title{Token Authentication UI}
\usage{
tokenAuthUI(id)
}
\arguments{
\item{id}{Character.Element ID}
}
\description{
Token Authentication UI. Adds a Login button which opens a
modal. Successful login will remove the button and replace it with the
username. Unsuccessful will open a different modal with an error.
}
\seealso{
\code{tokenAuthServer}, \code{tokenAuthModule}
}
