% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_integer_value}
\alias{validate_integer_value}
\title{Validate a value as an Integer}
\usage{
validate_integer_value(x)
}
\arguments{
\item{x}{A numeric or character value representing an integer.}
}
\value{
An integer value.
}
\description{
This function validates a value as an integer. If the value is numeric, it
checks that it is an integer with no decimal values. If the value is a
character, it checks that it is a numeric string with no decimal values.
If the value is not numeric or character, it throws an error.
}
\examples{
5L == validate_integer_value(5)   ## TRUE
5L == validate_integer_value("5") ## TRUE
validate_integer_value("5.0")     ## Error: value must be an integer with no decimal values.

}
