% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamsList.R
\name{ParamsList}
\alias{ParamsList}
\alias{push.ParamsList}
\alias{pop.ParamsList}
\alias{print.ParamsList}
\alias{clear.ParamsList}
\title{ParamsList S3 Class}
\usage{
ParamsList(params = list())

push.ParamsList(object, params)

pop.ParamsList(object, keys)

\method{print}{ParamsList}(object)

clear.ParamsList(object)
}
\arguments{
\item{params}{A list of key/value pairs to be added or updated in the ParamsList object.}

\item{object}{An object of class \code{ParamsList}.}

\item{keys}{A single key as a string or a list of keys.}
}
\value{
An object of class \code{ParamsList}.

The updated \code{ParamsList} object.

A list of key/value pairs that were removed.

The cleared \code{ParamsList} object.
}
\description{
An S3 class to manage a list of parameters with methods to push, pop,
print, and clear parameters.

Adds a list of key/value pairs to the ParamsList object. If a key already exists, its value is updated.

Removes and returns the value(s) associated with the provided key(s) from
the ParamsList object.

Prints the current state of the ParamsList object.

Clears all parameters from the ParamsList object.
}
\examples{
# Create an instance of ParamsList
params <- ParamsList()

# Push a single key/value pair
params <- push(params, "key1", "value1")
print(params)

# Push multiple key/value pairs
params <- push(params, list(key2 = "value2", key3 = "value3"))
print(params)

# Pop a single key/value pair
popped_value <- pop(params, "key1")
print(popped_value)
print(params)

# Pop multiple key/value pairs
popped_values <- pop(params, c("key2", "key3"))
print(popped_values)
print(params)

# Clear all parameters
params <- clear(params)
print(params)
}
\concept{ParamsList-methods}
