% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_response_utils.R
\name{process_tf_data}
\alias{process_tf_data}
\title{Process data and return a rank_response_df}
\usage{
process_tf_data(expr_df, binding_df, chipexo_df, cc_df)
}
\arguments{
\item{expr_df}{A data frame containing gene expression information}

\item{binding_df}{A data frame containing binding information}

\item{chipexo_df}{A data frame containing the chipexo information}

\item{cc_df}{A data frame containing chromatin capture information}
}
\value{
A rank_response_df data frame containing processed information
about TFs and their target genes
}
\description{
This function processes transcription factor (TF) data and returns a
rank_response_df, a data frame containing processed information about TFs
and their target genes.
}
\examples{
\dontrun{
# Assuming promoter_df, expr_df, binding_df, and cc_df are properly
# formatted data frames
rank_response_df <- process_tf_data(promoter_df, expr_df, binding_df, cc_df)
}

}
