% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankResponseAPI.R
\name{RankResponseAPI}
\alias{RankResponseAPI}
\title{RankResponseAPI R6 Class}
\value{
A RankResponseAPI object
}
\description{
Retrieves rank response data from the database.
}
\details{
An R6 class to interact with the Rank Response API.
}
\seealso{
Other API: 
\code{\link{AbstractAPI}}
}
\concept{API}
\section{Super class}{
\code{yeastCCDBShiny::AbstractAPI} -> \code{RankResponseAPI}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RankResponseAPI-new}{\code{RankResponseAPI$new()}}
\item \href{#method-RankResponseAPI-read}{\code{RankResponseAPI$read()}}
\item \href{#method-RankResponseAPI-clone}{\code{RankResponseAPI$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="yeastCCDBShiny" data-topic="AbstractAPI" data-id="create"><a href='../../yeastCCDBShiny/html/AbstractAPI.html#method-AbstractAPI-create'><code>yeastCCDBShiny::AbstractAPI$create()</code></a></span></li>
<li><span class="pkg-link" data-pkg="yeastCCDBShiny" data-topic="AbstractAPI" data-id="delete"><a href='../../yeastCCDBShiny/html/AbstractAPI.html#method-AbstractAPI-delete'><code>yeastCCDBShiny::AbstractAPI$delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="yeastCCDBShiny" data-topic="AbstractAPI" data-id="pop_params"><a href='../../yeastCCDBShiny/html/AbstractAPI.html#method-AbstractAPI-pop_params'><code>yeastCCDBShiny::AbstractAPI$pop_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="yeastCCDBShiny" data-topic="AbstractAPI" data-id="push_params"><a href='../../yeastCCDBShiny/html/AbstractAPI.html#method-AbstractAPI-push_params'><code>yeastCCDBShiny::AbstractAPI$push_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="yeastCCDBShiny" data-topic="AbstractAPI" data-id="update"><a href='../../yeastCCDBShiny/html/AbstractAPI.html#method-AbstractAPI-update'><code>yeastCCDBShiny::AbstractAPI$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RankResponseAPI-new"></a>}}
\if{latex}{\out{\hypertarget{method-RankResponseAPI-new}{}}}
\subsection{Method \code{new()}}{
Initializes the RankResponseAPI client with a URL, token, and optional parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RankResponseAPI$new(url = Sys.getenv("RANK_RESPONSE_URL"), ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RankResponseAPI-read"></a>}}
\if{latex}{\out{\hypertarget{method-RankResponseAPI-read}{}}}
\subsection{Method \code{read()}}{
Asynchronously retrieve rank response data from the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RankResponseAPI$read(callback = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{callback}}{A function to call when the data has been retrieved. The first argument of the callback should be \code{self}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A promise that resolves when the data has been retrieved and the metadata updated.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RankResponseAPI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RankResponseAPI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RankResponseAPI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
