% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamsList.R
\name{[[.ParamsList}
\alias{[[.ParamsList}
\title{Extract element from ParamsList using \code{[[}}
\usage{
\method{[[}{ParamsList}(x, i, ...)
}
\arguments{
\item{x}{An object of class \code{ParamsList}.}

\item{i}{The name of the parameter to extract.}

\item{...}{Additional arguments (not used).}
}
\value{
The value of the parameter.
}
\description{
Extract element from ParamsList using \code{[[}
}
