% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamsList.R
\name{[[<-.ParamsList}
\alias{[[<-.ParamsList}
\title{Assign element to ParamsList using \verb{[[<-}}
\usage{
\method{[[}{ParamsList}(x, i) <- value
}
\arguments{
\item{x}{An object of class \code{ParamsList}.}

\item{i}{The name of the parameter to assign.}

\item{value}{The value to assign to the parameter.}
}
\value{
The modified ParamsList object.
}
\description{
Assign element to ParamsList using \verb{[[<-}
}
\seealso{
Other ParamsList-methods: 
\code{\link{[.ParamsList}()},
\code{\link{[<-.ParamsList}()},
\code{\link{[[.ParamsList}()},
\code{\link{as.character.ParamsList}()},
\code{\link{as.list.ParamsList}()},
\code{\link{names.ParamsList}()},
\code{\link{print.ParamsList}()},
\code{\link{sort.ParamsList}()}
}
\concept{ParamsList-methods}
